requirements.txt
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
ipdb==0.13.9
requests==2.28.0
mysqlclient==2.1.1